ompio: fix abstraction
