+++ ompi/mca/common/ompio/common_ompio_print_queue.h
@@ -11,7 +11,7 @@
  * Copyright (c) 2004-2005 The Regents of the University of California.
  *                         All rights reserved.
  * Copyright (c) 2008-2016 University of Houston. All rights reserved.
- * Copyright (c) 2015      Research Organization for Information Science
+ * Copyright (c) 2015-2018 Research Organization for Information Science
  *                         and Technology (RIST). All rights reserved.
  * $COPYRIGHT$
  *
@@ -26,7 +26,6 @@
 
 #include "mpi.h"
 
-OMPI_DECLSPEC extern int mca_io_ompio_coll_timing_info;
 struct mca_io_ompio_file_t;
 
 #define MCA_COMMON_OMPIO_QUEUESIZE 2048
