+++ ompi/mca/io/ompio/Makefile.am
@@ -11,6 +11,8 @@
 #                         All rights reserved.
 # Copyright (c) 2008-2012 University of Houston. All rights reserved.
 # Copyright (c) 2016-2017 IBM Corporation.  All rights reserved.
+# Copyright (c) 2018      Research Organization for Information Science
+#                         and Technology (RIST).  All rights reserved.
 # $COPYRIGHT$
 #
 # Additional copyrights may follow
@@ -34,8 +36,7 @@ mcacomponentdir = $(ompilibdir)
 mcacomponent_LTLIBRARIES = $(component_install)
 mca_io_ompio_la_SOURCES = $(headers) $(sources)
 mca_io_ompio_la_LDFLAGS = -module -avoid-version
-mca_io_ompio_la_LIBADD = $(top_builddir)/ompi/lib@OMPI_LIBMPI_NAME@.la \
-	$(io_ompio_LIBS) \
+mca_io_ompio_la_LIBADD = $(io_ompio_LIBS) \
     $(OMPI_TOP_BUILDDIR)/ompi/mca/common/ompio/libmca_common_ompio.la
 
 noinst_LTLIBRARIES = $(component_noinst)
@@ -45,12 +46,10 @@ libmca_io_ompio_la_LDFLAGS = -module -avoid-version
 # Source files
 
 headers = \
-	io_ompio.h \
-	io_ompio_aggregators.h
+	io_ompio.h
 
 sources = \
         io_ompio.c \
-        io_ompio_aggregators.c \
         io_ompio_component.c \
         io_ompio_module.c \
         io_ompio_file_set_view.c \
